리팩토링
Refactor: 코드 리팩토링
각 파일들이 속한 폴더를 myPuppy로 변경
