+++ src/controller/my_puppy/puppy_update.js
@@ -2,6 +2,7 @@ import query from "../../mysql/index.js";
 import sql from "../../mysql/sql.js";
 
 export const update = async(req, res) => {
+    
     //res.send('내 강아지 메인 페이지');
     res.send('/my-puppy/update-profile 라우트');
 }
